fastercnn-pytorch-training-pipeline/requirements.txt
# Base-------------------------------------
albumentations>=1.1.0
ipython
jupyter
matplotlib
opencv-python>=4.1.1.26
opencv-python-headless>=4.1.1.26
Pillow
PyYAML
scikit-image
scikit-learn
scipy
torch>=1.12.0
torchvision>=0.13.0
numpy
protobuf<=3.20.1
pandas
tqdm

# Logging----------------------------------
# wandb
tensorboard

# Model summary----------------------------
torchinfo

# Extras-----------------------------------
pycocotools>=2.0.2
setuptools==59.5.0
torchmetrics # Evaluation

# Transformer based models.
vision_transformers

# Export-----------------------------------
# onnxruntime
# onnx
# onnxruntime-gpu